[ao][work-in-progress] Implemented InputWeightEqualization Detector
Summary: This adds the implementation for the InputWeightEqualization
detector. This includes both the implementation and the relavent test
cases. This detector is meant to be added to initialize a ModelReport
instance and it will keep track of the necessary statistics to decide if
for certain layers of interest (linear and conv for now), it makes sense
to use input weight equalization and gives the suggestion to the user.

This code is still work in progress and has so far implemented finding
detector insert points, and is still needing generating report
funcitonality.

Test Plan: python test/test_quantization.py TestFxDetectInputWeightEqualization

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 78c3fb3aa6e9de32c34171cf088ab1297e722b99
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79962